fix(api): organize logic in different files
Ref. metriport/metriport-internal#1333
