+++ packages/api/src/domain/base-domain.ts
@@ -11,9 +11,3 @@ export interface BaseDomain extends BaseDomainCreate {
 export interface BaseDomainSoftDelete extends BaseDomain {
   deletedAt?: Date;
 }
-
-export interface BaseResultDomain extends BaseDomainCreate {
-  requestId: string;
-  status: string;
-  createdAt: Date;
-}
