+++ packages/api/src/domain/medical/ihe-result.ts
@@ -1,5 +1,16 @@
-import { BaseResultDomain } from "../base-domain";
+import { PatientDiscoveryResponse, DocumentQueryResponse } from "@metriport/ihe-gateway-sdk";
+import { BaseDomainCreate } from "../base-domain";
 
-export interface IHEResult<Response> extends BaseResultDomain {
-  data: Response;
+export interface BaseResultDomain extends BaseDomainCreate {
+  requestId: string;
+  status: string;
+  createdAt: Date;
+}
+
+export interface PatientDiscoveryResult extends BaseResultDomain {
+  data: PatientDiscoveryResponse;
+}
+
+export interface DocumentQueryResult extends BaseResultDomain {
+  data: DocumentQueryResponse;
 }
